internal: move types-all to lib/types
