--- generators/base-application/application-options-all.d.ts
@@ -1,30 +1,0 @@
-/* eslint-disable @typescript-eslint/consistent-type-imports */
-import type { Simplify } from 'type-fest';
-import type { ExportGeneratorOptionsFromCommand } from '../../lib/command/types.js';
-import type { Options as ClientOptions } from '../client/types.d.ts';
-import type { Options as JavascriptOptions } from '../javascript/types.d.ts';
-import type { Options as SpringCacheOptions } from '../spring-cache/index.js';
-import type { Options as SpringCloudStreanOptions } from '../spring-cloud-stream/index.js';
-
-export type ApplicationOptions = Simplify<
-  JavascriptOptions &
-    ClientOptions &
-    SpringCacheOptions &
-    SpringCloudStreanOptions &
-    ExportGeneratorOptionsFromCommand<typeof import('../app/command.ts').default> &
-    ExportGeneratorOptionsFromCommand<typeof import('../base/command.ts').default> &
-    ExportGeneratorOptionsFromCommand<typeof import('../bootstrap-application-base/command.ts').default> &
-    ExportGeneratorOptionsFromCommand<typeof import('../client/command.ts').default> &
-    ExportGeneratorOptionsFromCommand<typeof import('../git/command.ts').default> &
-    ExportGeneratorOptionsFromCommand<typeof import('../java/generators/bootstrap/command.ts').default> &
-    ExportGeneratorOptionsFromCommand<typeof import('../java/generators/build-tool/command.ts').default> &
-    ExportGeneratorOptionsFromCommand<typeof import('../java/generators/graalvm/command.ts').default> &
-    ExportGeneratorOptionsFromCommand<typeof import('../jdl/command.ts').default> &
-    ExportGeneratorOptionsFromCommand<typeof import('../languages/command.ts').default> &
-    ExportGeneratorOptionsFromCommand<typeof import('../liquibase/command.ts').default> &
-    ExportGeneratorOptionsFromCommand<typeof import('../project-name/command.ts').default> &
-    ExportGeneratorOptionsFromCommand<typeof import('../server/command.ts').default> &
-    ExportGeneratorOptionsFromCommand<typeof import('../spring-boot/command.ts').default> &
-    ExportGeneratorOptionsFromCommand<typeof import('../spring-cloud/generators/gateway/command.ts').default> &
-    ExportGeneratorOptionsFromCommand<typeof import('../spring-data-relational/command.ts').default>
->;
