+++ generators/base-application/generator.ts
@@ -33,7 +33,7 @@ import type { GenericTaskGroup } from '../base-core/types.js';
 import type { Entity as BaseEntity } from '../../lib/jhipster/types/entity.js';
 import { getConfigWithDefaults } from '../../lib/jhipster/default-application-options.js';
 import { BOOTSTRAP_APPLICATION } from '../base-simple-application/priorities.js';
-import type { ApplicationConfiguration } from './application-config-all.js';
+import type { ApplicationConfiguration } from '../../lib/types/application-config-all.js';
 import type {
   ConfiguringEachEntityTaskParam,
   TaskTypes as DefaultTasks,
@@ -42,8 +42,8 @@ import type {
   PreparingEachEntityRelationshipTaskParam,
   PreparingEachEntityTaskParam,
 } from './tasks.js';
-import type { ApplicationAll } from './application-properties-all.js';
-import type { SourceAll } from './source-all.js';
+import type { ApplicationAll } from '../../lib/types/application-properties-all.js';
+import type { SourceAll } from '../../lib/types/source-all.js';
 import { CONTEXT_DATA_APPLICATION_ENTITIES_KEY, getEntitiesFromDir } from './support/index.js';
 import { CUSTOM_PRIORITIES, PRIORITY_NAMES, QUEUES } from './priorities.js';
 import type {
