+++ generators/base-application/support/task-type-inference.ts
@@ -24,9 +24,9 @@ import type {
   Source as BaseApplicationSource,
 } from '../types.js';
 
-import type { ApplicationAll } from '../application-properties-all.js';
-import type { EntityAll } from '../entity-all.js';
-import type { SourceAll } from '../source-all.js';
+import type { ApplicationAll } from '../../../lib/types/application-properties-all.js';
+import type { EntityAll } from '../../../lib/types/entity-all.js';
+import type { SourceAll } from '../../../lib/types/source-all.js';
 
 export function asWriteFilesSection<Data = ApplicationAll<EntityAll>>(section: WriteFileSection<Data>) {
   return section;
