+++ generators/java/support/files.ts
@@ -20,8 +20,8 @@
 import type { WriteFileBlock } from '../../base-core/api.js';
 import type CoreGenerator from '../../base-core/generator.js';
 import { SERVER_MAIN_RES_DIR, SERVER_MAIN_SRC_DIR, SERVER_TEST_RES_DIR, SERVER_TEST_SRC_DIR } from '../../generator-constants.js';
-import type { EntityAll } from '../../base-application/entity-all.js';
-import type { ApplicationAll } from '../../base-application/application-properties-all.js';
+import type { EntityAll } from '../../../lib/types/entity-all.js';
+import type { ApplicationAll } from '../../../lib/types/application-properties-all.js';
 
 export const replaceEntityFilePathVariables = (data: any, filePath: string) => {
   filePath = filePath
