+++ lib/jdl/core/types/json-config.d.ts
@@ -1,4 +1,4 @@
-import type { ApplicationConfiguration, YoRcContent } from '../../../../generators/base-application/application-config-all.js';
+import type { ApplicationConfiguration, YoRcContent } from '../../../types/application-config-all.js';
 import type { Entity } from '../../../jhipster/types/entity.js';
 import type { Field } from '../../../jhipster/types/field.js';
 import type { Relationship } from '../../../jhipster/types/relationship.js';
