+++ lib/jhipster/default-application-options.ts
@@ -17,7 +17,7 @@
  * limitations under the License.
  */
 
-import type { ApplicationConfiguration } from '../../generators/base-application/application-config-all.js';
+import type { ApplicationConfiguration } from '../types/application-config-all.js';
 import applicationTypes from './application-types.js';
 import authenticationTypes from './authentication-types.js';
 import databaseTypes from './database-types.js';
