+++ lib/types/field-all.d.ts
@@ -1,0 +1,5 @@
+import type { Field as SpringDataRelationalField } from '../../generators/spring-data-relational/index.js';
+import type { Field as LiquibaseField } from '../../generators/liquibase/index.js';
+import type { Field as ClientField } from '../../generators/client/index.js';
+
+export type FieldAll = SpringDataRelationalField & LiquibaseField & ClientField;
