+++ lib/utils/yo-rc.ts
@@ -1,6 +1,6 @@
 import { readFileSync } from 'fs';
 import { join } from 'path';
-import type { ApplicationConfiguration, YoRcContent } from '../../generators/base-application/application-config-all.js';
+import type { ApplicationConfiguration, YoRcContent } from '../types/application-config-all.js';
 import type { Entity } from '../jhipster/types/entity.js';
 import { upperFirst } from '../jdl/core/utils/string-utils.js';
